requirements.txt
fastapi
uvicorn
google-genai
google-adk
mcp
python-dotenv
pydantic
